+++ src/lib/favorites-controller.ts
@@ -1,8 +1,9 @@
 import type {TabInfo} from '_bg/getTabData'
+import {userController} from '@/lib/user-controller'
 
 export type Favorite = TabInfo
 
-export const favoritesManager = {
+export const favoritesController = {
   getFavorites(): Favorite[] {
     try {
       return JSON.parse(localStorage.getItem('favorites') ?? '[]')
@@ -19,25 +20,31 @@ export const favoritesManager = {
     return this.getFavorites().some((fav) => fav.url === url)
   },
 
-  addFavorite(data: Favorite): void {
+  async addFavorite(data: Favorite): Promise<void> {
     if (this.isFavorite(data.url)) return
 
     const favorites = [...this.getFavorites(), data]
     localStorage.setItem('favorites', JSON.stringify(favorites))
+
+    userController.updateFavorites()
+    await userController.sync()
   },
 
-  removeFavorite(url: string): void {
+  async removeFavorite(url: string): Promise<void> {
     const favorites = this.getFavorites().filter((fav) => fav.url !== url)
     localStorage.setItem('favorites', JSON.stringify(favorites))
+
+    userController.updateFavorites()
+    await userController.sync()
   },
 
-  toggleFavorite(data: Favorite): boolean {
+  async toggleFavorite(data: Favorite): Promise<boolean> {
     if (this.isFavorite(data.url)) {
-      this.removeFavorite(data.url)
+      await this.removeFavorite(data.url)
       return false
     }
 
-    this.addFavorite(data)
+    await this.addFavorite(data)
     return true
   },
 }