+++ src/modules/popup/details/Details.tsx
@@ -1,4 +1,8 @@
 import type {Units} from '_scripts/index'
+
+import {developerController} from '@/lib/developer-controller'
+import {userController} from '@/lib/user-controller'
+
 import {useEffect, useState} from 'react'
 
 import Layout from '~/Global/Layout'
@@ -18,6 +22,9 @@ export function Details() {
       })
     })
   }, [])
+
+  developerController.initControls()
+  userController.init()
 
   return (
     <Layout className="space-y-2.5">